% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_tecan}
\alias{read_tecan}
\title{Reads one or all sheets of a Tecan Plate reader Infinite 200 excel file.
Automatically determines the table format and the number of wells that were recorded.
Reads all absorbance values accordingly and records all sheet names and file names.
Ignores but warns about empty or malformatted sheets.}
\usage{
read_tecan(xlsx_path, sheet = "all")
}
\arguments{
\item{xlsx_path}{Path to the Tecan file.}

\item{sheet}{Name of the sheet to read (default: "all").}
}
\value{
A tibble in long format.
}
\description{
Reads one or all sheets of a Tecan Plate reader Infinite 200 excel file.
Automatically determines the table format and the number of wells that were recorded.
Reads all absorbance values accordingly and records all sheet names and file names.
Ignores but warns about empty or malformatted sheets.
}
\examples{
read_tecan("tecan_example.xlsx")
}
